requirements.txt
# requirements.txt
azure-functions
azure-storage-blob
psycopg2-binary